scripts/requirements.txt
beautifulsoup4
openai
python-dotenv
requests
